requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pynacl==1.5.0
pydantic==2.8.2
